move lora forward
